V1, V3의 SemanticColor를 공용으로 사용할 수 있는 Protocol 추가
