+++ Sources/BezierSwift/Foundation/Color/Themeable.swift
@@ -20,7 +20,7 @@ extension Themeable {
     }
   }
   
-  public func palette(_ semanticColor: SemanticColorToken, isInverted: Bool = false) -> Color {
+  public func palette(_ semanticColor: SemanticColorProtocol, isInverted: Bool = false) -> Color {
     switch self.colorScheme {
     case .light: return isInverted ? semanticColor.dark.color : semanticColor.light.color
     case .dark: return isInverted ? semanticColor.light.color : semanticColor.dark.color
